Remove collections.old package completely (#180)
