+++ base/src/main/java/org/eclipse/serializer/collections/KeyValueFlatCollector.java
@@ -1,4 +1,4 @@
-package org.eclipse.serializer.collections.old;
+package org.eclipse.serializer.collections;
 
 /*-
  * #%L
