--- base/src/main/java/org/eclipse/serializer/collections/old/OldCollection.java
@@ -1,24 +1,0 @@
-package org.eclipse.serializer.collections.old;
-
-/*-
- * #%L
- * Eclipse Serializer Base
- * %%
- * Copyright (C) 2023 MicroStream Software
- * %%
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
- * #L%
- */
-
-import java.util.Collection;
-
-import org.eclipse.serializer.collections.types.XGettingCollection;
-
-public interface OldCollection<E> extends Collection<E>
-{
-	public XGettingCollection<E> parent();
-}
