+++ persistence/binary/src/main/java/org/eclipse/serializer/persistence/binary/java/util/concurrent/BinaryHandlerConcurrentSkipListMap.java
@@ -17,8 +17,8 @@ package org.eclipse.serializer.persistence.binary.java.util.concurrent;
 import java.util.Comparator;
 import java.util.concurrent.ConcurrentSkipListMap;
 
-import org.eclipse.serializer.collections.old.KeyValueFlatCollector;
-import org.eclipse.serializer.collections.old.OldCollections;
+import org.eclipse.serializer.collections.KeyValueFlatCollector;
+import org.eclipse.serializer.persistence.binary.java.util.AbstractBinaryHandlerMap;
 import org.eclipse.serializer.persistence.binary.types.AbstractBinaryHandlerCustomCollection;
 import org.eclipse.serializer.persistence.binary.types.Binary;
 import org.eclipse.serializer.persistence.types.Persistence;
@@ -152,7 +152,7 @@ extends AbstractBinaryHandlerCustomCollection<ConcurrentSkipListMap<?, ?>>
 		final PersistenceLoadHandler      handler
 	)
 	{
-		OldCollections.populateMapFromHelperArray(instance, data.getHelper(instance));
+		AbstractBinaryHandlerMap.populateMapFromHelperArray(instance, data.getHelper(instance));
 	}
 
 	@Override
